Rename timecode format 29fps to 30-drop-frame
This matches the spec ("-29 corresponds to 30 drop frame").

This commit also introduces debug asserts in the implementation of
From<Division> for Vec<u8>, requiring that the marker bit mask is
appropriately set. This is guaranteed by the fact that all the
timecode format values are negative i8, and the number of ticks
per quarter note never exceeds 2^16 - 1.